azure-openai-workload-identity/requirements.txt
azure-identity==1.15.0
openai==1.1.1